calcEffects {FSDM}	R Documentation

Calculate updated concept values by applying model relations to a set of concept values and changes to those values.

Description

calcRelationEffects updates concept values by applying relations to a set of concept values and changes to those concept values.

Usage

calcEffects(V_Cn, R_Cn, M)

Arguments

V_Cn
a numeric vector of values in the interval (0, 100] identifying the relative value of each concept.

R_Cn
a numeric vector of values identifying the proportional change in each value.

M
a list containing the FSDM model components

Details

Since most models include cycles, the model calculations must be iterated until the concept values change very little between iterations or until a maximum number of iterations has been completed. In each iteration, the proportional changes to concept values from their starting values are updated by applying the calcPropChange function to calculate the proportional change due to the proportional change in each of the anterior concepts affecting the concept and then adding up the proportional change differences. This is done for each of the concepts in the model. For example, if a posterior concept has 2 anterior concepts affecting it and if the change ratio from one of them is 1.2 (i.e. the posterior concept is 1.2 times its previous value) and the change ratio from the other is 1.3. Then the sum of the proportional change differences is 0.5 and the total proportional change is 1.5.

Value

a list having 3 named components. Ratio is the ratio of change of the 'independent' concepts. Values is a numeric vector of the value of concepts after 'independent' concepts have been incremented. MaxValueChg is a number identifying the maximum percentage change in the values of concepts between starting and ending values.

[Package FSDM version 0.71 Index]