copyDemoModels {FSDM}	R Documentation

Copy demo model files from FSDM package

Description

copyDemoModels copies demo FSDM model files to specified folder

Usage

copyDemoModels(ToFolder)

Arguments

ToFolder
a string identifying the path to the folder on the users computer that the demo files are to be copied to.

Details

This function copies demo FSDM model files from the FSDM package to a specified folder. If a folder named 'models' is located in the specified folder, the function copies the demo models to that folder. If the folder does not exist, then it is created and the demo models copied to it.

[Package FSDM version 0.71 Index]