createFuzzyModel {FSDM}	R Documentation

Create fuzzy model

Description

createFuzzyModel creates the representation of a FSDM model as an R object from JSON-formatted text files.

Usage

createFuzzyModel(Dir, Vals = c(VL = 0.01, L = 0.2, ML = 0.35, M = 0.5, MH
  = 0.67, H = 0.85, VH = 1), Signs = c(Positive = 1, Negative = -1))

Arguments

Dir
a string identifying the path to the directory where the JSON-formatted text files that specify a model are located.

Vals
a named numeric vector that relates linguistic relationships to numeric values.

Signs
a named numeric vector which associates values with descriptions of relationship signs. Default is c(Positive = 1, Negative = -1).

Details

This function reads in JSON-formatted text files which contains all of the information needed to specify a FSDM model and makes an object (a list) which contains components in the data structures needed to apply the FSDM functions to compute an output given a scenario which specifies initial conditions.

Value

A list containing the following components: Cn = a string vector containing the names of the model concepts; Group = a named string vector containing the group name for each concept; Relates = a numeric matrix whose dimensions are equal to the number of concepts and values are the numeric weights in the model; Labels = a string matrix, with the same dimensions as Relates, which contains the fuzzy relationships between concepts (e.g. low, medium, high); ValueRange = a data frame which provides the minimum and maximum values of each concept.

[Package FSDM version 0.71 Index]