createFuzzyScenario {FSDM}	R Documentation

Create a scenario object to be modeled

Description

createFuzzyScenario loads a scenario file into object used in model application

Usage

createFuzzyScenario(Dir, M, OpRange = c(0.01, 99.99))

Arguments

Dir
the a string identifying the path to the scenario directory.

M
the FSDM model object that is created by the 'createFuzzyModel' function.

OpRange
a 2-element vector identifying the operating range of the model. The default values are c(0.01, 99.99), just short of 0 and 100 to avoid values that either result in no change or infinite change.

Details

This function reads a JSON formatted text file which describes a scenario. A scenario is defined by the starting values of all concepts and starting changes in one or more concepts.

Value

a list containing the following components: Cn a vector of concept variable names in the same order as the list in the fuzzy model. StartValues a numeric vector of concept starting values scaled to the range of 0 to 100 and in the order of Cn. ChangeTo a numeric vector of concept starting changes scaled to the operating range and in the order of Cn.

[Package FSDM version 0.71 Index]