formatConceptTable {FSDM}	R Documentation

Formats a concept table to be displayed in the GUI.

Description

formatConceptTable formats a concept data frame to be displayed as a table in the Logic Laboratory GUI.

Usage

formatConceptTable(Concepts_df)

Arguments

Concepts_df
a data frame containing the concepts data.

Details

The GUI summarizes information about model concepts in a table. Not all of the concept data needs to be shown and some of the data is difficult to show in table form. This function extracts and formats the concept data that is to be displayed in a table.

Value

a data frame containing the concepts data to be shown in a table.

[Package FSDM version 0.71 Index]