formatOutputData {FSDM}	R Documentation

Create data frame of selected scenarios and concepts to plot and save

Description

formatOutputData makes a data frame of the summary results for selected scenarios and selected concepts.

Usage

formatOutputData(ModelDir, ModelName, Sc, Vn)

Arguments

ModelDir
a string identifying the path to the models folder in which

ModelName
a string representation of the model name.

Sc
a vector of the names of scenarios to include.

Vn
the variable names for the concepts to include.

Details

This function creates a data frame of model results for selected scenarios and selected concepts. The data frame is in 'flat' format where values are in one column and the corresponding concept names, scenario names, and iterations are in separate columns.

Value

A data frame having columns identifying the scenario, concept, iteration, scaled values, and rescaled values.

[Package FSDM version 0.71 Index]