getEffects {FSDM}	R Documentation

Get data on effect relationships of a causal concept

Description

getEffects extracts data on all the effect relationships of an identified causal concept.

Usage

getEffects(Model_ls, ConceptName)

Arguments

Model_ls
a list that includes components for model concepts and relations

ConceptName
a string containing the name of the causal concept

Details

The function extracts the data on the effect relationships of an identified causal concept. This includes the names of the affected concepts and the respective effect directions, magnitudes, and descriptions.

Value

a data frame containing the names, directions, magnitudes, and descriptions of all the effects or NULL if there are no effects.

[Package FSDM version 0.71 Index]