idScenWithOutputs {FSDM}	R Documentation

Identify scenarios that have model outputs

Description

idScenWithOutputs returns the names of scenarios that have model run outputs.

Usage

idScenWithOutputs(ModelsDir, ModelName)

Arguments

ModelsDir
a string identifying the path to the models folder in which

ModelName
a string representation of the model name.

Details

This function takes the name of a model and returns a vector of the names of all the model scenarios that have model outputs.

Value

a string vector of the names of scenarios having model outputs.

[Package FSDM version 0.71 Index]