initRelationsEntry {FSDM}	R Documentation

Initialize a initial relations entry for a new concept

Description

initRelationsEntry creates a initial relations entry for concept that is being added to the model

Usage

initRelationsEntry(VarName)

Arguments

VarName
a string representation of the concept variable name

Details

This function creates an initial relations entry data for a concept. This entry data has the concept variable name and empty fields for all other data items. The server script calls this function when a concept is created and adds the resulting entry to the relations table.

Value

a data frame which includes all the mandatory relations fields with the name field populated with the concept variable name and all other fields populated with empty fields

[Package FSDM version 0.71 Index]