initializeCopyModel {FSDM}	R Documentation

Initialize a new model by copying an existing model.

Description

initializeCopyModel initializes a new model by creating a directory with the model name and copying the contents of an existing model into that directory. Creates and saves a model status list.

Usage

initializeCopyModel(ModelsDir, ModelName, CopyModelName, Author,
  CopyScenarios = FALSE)

Arguments

ModelsDir
a string identifying the path to the models folder in which

ModelName
a string identifying the model name.

CopyModelName
a string identifying the name of the model to copy.

Author
a string identifying the name of the author.

CopyScenarios
a logical to determine whether to copy the model scenarios.

Details

This function initializes a new model with the given model name from an existing model. It does this by creating a directory with the model name and copying the model files for an existing model into it. It also copies and updates the model status list by identifying the new model name, the parent model (i.e. copied model) name, time and date of creation, edit and attribution history, and model notes. The function can be used to copy only the model or all the scenarios as well as the model

Value

a model status list for the model which includes information about the copied model.

[Package FSDM version 0.71 Index]