initializeCopyScenario {FSDM}	R Documentation

Initialize a new scenario by copying an existing scenario.

Description

initializeCopyScenario initializes a new scenario by creating a directory with the scenario name and copying the contents of an existing scenario into that directory.

Usage

initializeCopyScenario(ModelsDir, ModelName, ConceptVars_, ScenarioName,
  CopyScenarioName, NumIncr)

Arguments

ModelsDir
a string identifying the path to the models folder in which the model is located.

ModelName
a string representation of the model name.

ConceptVars_
a string vector of the model concept variable names.

ScenarioName
a string representation of the new scenario name.

CopyScenarioName
a string representation of the name of the scenario to copy.

NumIncr
an integer identifying the number of growth increments for the scenario.

Details

This function initializes a new scenario with the given scenario name from an existing scenario. It does this by creating a directory with the scenario name and copying the scenario files for an existing scenario into it. It creates a new scenario status list which identifies the name of the new scenario and the name of the parent scenario it is a copy of. It also identifies the number of growth increments in the new scenario.

Value

a list containing a status list and a scenario values data frame.

[Package FSDM version 0.71 Index]