initializeNewModel {FSDM}	R Documentation

Initialize a new model.

Description

initializeNewModel initializes a new model by creating a directory with the model name. Creates and saves a model status list.

Usage

initializeNewModel(ModelsDir, ModelName, Author)

Arguments

ModelsDir
a string identifying the path to the models folder in which the model is located.

ModelName
a string identifying the model name.

Author
a string identifying the author's name.

Details

This function initializes a new model with the given model name. It does this by creating a directory with the model name and a list to store the model status. This list is saved in the model directory in JSON format and is also returned by the function. The status list contains the name of the model, the parent (none), the date and time is was created, and the date and time it was last edited (same as creation time).

Value

a list containing values for name, parent, created, and lastedit.

[Package FSDM version 0.71 Index]