initializeNewScenario {FSDM}	R Documentation

Initialize a new scenario.

Description

initializeNewScenario initializes a new scenario by creating a directory with the scenario name and returns a list containing scenario status, scenario values, and number of growth increments components.

Usage

initializeNewScenario(ModelsDir, ModelName, ScenarioName, Concepts_df,
  NumIncr)

Arguments

ModelsDir
a string identifying the path to the models folder in which

ModelName
a string representation of the model name.

ScenarioName
a string representation of the scenario name.

Concepts_df
a data frame containing the concept data.

NumIncr
an integer identifying the number of growth increments for the scenario.

Details

This function initializes a new scenario with the given scenario name. It does this by creating a directory with the scenario name and a data frame to store the scenario data. This list is returned and the components are used in the scenario reactive object. The components are also saved in the scenario directory in JSON format.

Value

a list containing a status list and a scenario values data frame.

[Package FSDM version 0.71 Index]