listScenarios {FSDM}	R Documentation

Creates a list of scenarios and whether validated.

Description

listScenarios creates a list of scenarios and identifies which have been validated.

Usage

listScenarios(ModelsDir, ModelName)

Arguments

ModelsDir
a string identifying the path to the models folder in which the operating model is located.

ModelName
the name of the model

Details

This function creates a list of scenarios for a model and checks whether each scenario has been validated. Scenarios are identified as being validated if they were successfully validated and if the validation time stamp is later than the 'lastedit' time stamp for the model and the 'lastedit' time stamp for the scenario. The returned list has two components. The first is a vector containing the names that have been successfully validated. The second is a vector of the names that have not been validated; either because validation was not successful or because changes were made to the model or the scenario after validation was done.

Value

a list having four components, a vector of the names of scenarios that were properly validated, a vector of the names of scenarios that were not validated, a vector of the names of all scenarios, and a vector of the names of scenarios that have outputs.

[Package FSDM version 0.71 Index]