loadModelRelations {FSDM}	R Documentation

Load the relations file for a model.

Description

loadModelRelations reads the file that contains model relations information and returns a data frame representation.

Usage

loadModelRelations(ModelsDir, ModelName)

Arguments

ModelsDir
a string identifying the path to the models folder in which

ModelName
a string representation of the model name.

Details

This function reads the model relations file for a specified model and returns a data frame containing the information.

Value

a data frame containing the model relations information.

[Package FSDM version 0.71 Index]