loadModelStatus {FSDM}	R Documentation

Load the model status file for a model.

Description

loadModelStatus reads a model status file and returns a list containing the status information.

Usage

loadModelStatus(ModelsDir, ModelName, Author = NULL)

Arguments

ModelsDir
a string identifying the path to the models folder in which

ModelName
a string representation of the model name.

Author
a string identifying the author's name

Details

This function reads the model status JSON file for a specified model and creates a list containing the model status information.

Value

a list containing values for name, parent, created, and lastedit.

[Package FSDM version 0.71 Index]