loadScenario {FSDM}	R Documentation

Loads the files for a scenario.

Description

loadScenario reads the files that contain scenario information and returns a list containing scenario status, values, and increments information.

Usage

loadScenario(ModelsDir, ModelName, ConceptVars_, ScenarioFileName)

Arguments

ModelsDir
a string identifying the path to the models folder in which

ModelName
a string representation of the model name.

ConceptVars_
a string vector of the model concept variable names.

ScenarioFileName
a string representation of the scenario name.

Details

This function reads the scenario status file and scenario file for a specified scenario and returns a list whose components are a list containing the scenario status information, a data frame containing the scenario values information, and an integer identifying the number of growth increments.

Value

a list containing a status list, a scenario values data frame, and the number of growth increments for the scenario.

[Package FSDM version 0.71 Index]