makeAdjacencyMatrix {FSDM}	R Documentation

Creates an adjacency matrix from an FSDM relations list

Description

makeAdjacencyMatrix creates an adjacency matrix from an FSDM relations list

Usage

makeAdjacencyMatrix(Relations_ls, Type = "Logical")

Arguments

Relations_ls
a list of FSDM relations

Type
a string identifying the type of matrix to create. "Logical" produces a matrix where the existence of a relationship is noted with TRUE and FALSE. "Values" produces a list of two matrices where one matrix identifies the causal weitht of each relationship and a second matrix identifies the causal direction of each relationship.

Details

This function creates an adjacency matrix from an FSDM relations list. The adjacency matrix is a square matrix with as many rows and columns as the number of concepts in the model. The rows and columns are named with the concept variable names. The rows represent the causal side of the relationship and the columns represent the affected side. The values in the matrix are logicals with TRUE meaning that a relationship exists and FALSE meaning that it does not.

Value

a matrix of logical values or string values

[Package FSDM version 0.71 Index]