makeDot {FSDM}	R Documentation

Create a DOT file for GraphViz

Description

makeDotFile create and save a dot file to be displayed by GraphViz

Usage

makeDot(Relations_ls, Concepts_df, RowGroup = "All", ColGroup = "All",
  orientation = "Portrait", rankdir = "Top-to-Bottom", shape = "box",
  Show = "label")

Arguments

Relations_ls
a list of model relations.

Concepts_df
a data frame with model concepts.

RowGroup
a string identifying the names of the name of the group that selects rows from the relationship table to plot. The default 'All' selects all the rows.

ColGroup
a string identifying the name of the group that selects columns from the relationship table to plot. The default 'All' selects all the columns.

orientation
a string identifying the GraphViz layout orientation ('Portrait' or 'Landscape').

rankdir
a string identifying the graph orientation: 'TB' for top to bottom, 'LR' for left to right.

shape
a string identifying the shape of the graph nodes (e.g. 'box').

Show
a string identifying how to label the edges. The default value "label" results in showing the fuzzy label (e.g. VL, L, M, H, VH). The alternative, "value", results in showing the equivalent numeric value.

Details

This function writes out a dot file to be rendered using GraphViz.

Value

A string specification of a DOT.

[Package FSDM version 0.71 Index]