mapRelations {FSDM}	R Documentation

Plot relationships

Description

mapRelations displays a plot of concept relations, highlighting the relations from or to a highlighted concept.

Usage

mapRelations(Model_ls, FromConcept = NULL, FromGroup = "All",
  ToGroup = "All")

Arguments

Model_ls
a list that includes components for model concepts and relations

FromConcept
the name of a selected causal concept or NULL

FromGroup
the name of the group of the causal concepts to display or NULL

ToGroup
the name of the group of the receiving concepts to display or NULL

Details

The function maps relationships between concepts by plotting concepts in two parallel vertical lines and then showing relationships between concepts by drawing lines between the related concepts. The selected concept and related concepts are placed at the top of the plot and the lines are highlighted.

Value

a logical value identifying whether the plot can be created

[Package FSDM version 0.71 Index]