rescale {FSDM}	R Documentation

Rescale value

Description

rescale rescales a value from a specified input range to a specified output range.

Usage

rescale(Value, FromRange, ToRange)

Arguments

Value
a numeric value to be rescaled from the input range to the output range.

FromRange
a numeric vector of length 2 in which the first value is the minimum value in the range and the second value is the maximum value in the range.

ToRange
a numeric vector of length 2 in which the first value is the minimum value in the range and the second value is the maximum value in the range.

Details

This function rescales a value from a specified input range to a specified output range. The default output range is 0 to 100 because that is the range used to represent concepts in FSDM models.

Value

A numeric value in the output range.

[Package FSDM version 0.71 Index]