runLogicLab {FSDM}	R Documentation

Run the Logic Laboratory

Description

runLogicLab runs the Logic Laboratory, the GUI that can be used to build models and scenarios, and to run them and examine the results

Usage

runLogicLab()

Details

The graphical user interface for building and running FSDM models/scenarios is called the Logic Laboratory. This is a Shiny application which resides in the 'inst/logic-lab' directory of the source package and the 'logic-lab' directory of the installed package. This function starts the Logic Laboratory application. The function has no parameters and returns no results.

[Package FSDM version 0.71 Index]