saveModel {FSDM}	R Documentation

Saves all the model components as JSON files.

Description

saveModel saves the model status, model concepts, and model relations as JSON files.

Usage

saveModel(ModelsDir, ModelData)

Arguments

ModelsDir
a string identifying the path to the models folder in which

ModelData
a model.

Details

Models are composed of 3 objects: a model status list, a model concepts data frame, and a model relations data frame. This function saves these objects as JSON-formatted files.

Value

no return value. Has side effect of saving the model status list, model concepts data frame, and model relations data frame as JSON-formatted files.

[Package FSDM version 0.71 Index]