saveScenario {FSDM}	R Documentation

Saves all the scenario components as JSON files.

Description

saveScenario saves the scenario status and values as JSON files.

Usage

saveScenario(ModelsDir, ScenarioData)

Arguments

ModelsDir
a string identifying the path to the models folder in which

ScenarioData
a list having status and values components.

Details

This function saves the scenario status and values as JSON files in the scenario directory for the model. The status information is saved in the status.json file and the scenario values are saved in the scenarios.json file.

Value

no return value. Has side effect of saving the scenario status list and values data frame.

[Package FSDM version 0.71 Index]