validateScenario {FSDM}	R Documentation

Validates scenario with model.

Description

validateScenario validates the scenario with the model.

Usage

validateScenario(Values_df, Concepts_df)

Arguments

Values_df
a data frame containing the scenario values

Concepts_df
a data frame containing the model concepts

Details

This function compares the values for the scenario with the model and determines whether the scenario has values for all concepts, whether the values are consistent with the value ranges for the concepts, and whether the starting changes are consistent with the ranges and don't vary too many.

Value

a list having three components, a logical identifying whether the scenario validates, a list of all validation errors, and a character string with the time stamp.

[Package FSDM version 0.71 Index]